Move Glue tests to dedicated package in Iceberg
To run all Glue tests easily on IDE.
